+++ src/main/java/hansung/mannayo/mannayoserverapplication/Repository/BlockRepository.java
@@ -3,5 +3,5 @@ package hansung.mannayo.mannayoserverapplication.Repository;
 import hansung.mannayo.mannayoserverapplication.Model.Entity.Block;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BlockRepository extends JpaRepository<Block,Integer> {
+public interface BlockRepository extends JpaRepository<Block,Long> {
 }
