+++ src/main/java/hansung/mannayo/mannayoserverapplication/Repository/ReportRepository.java
@@ -3,5 +3,5 @@ package hansung.mannayo.mannayoserverapplication.Repository;
 import hansung.mannayo.mannayoserverapplication.Model.Entity.Report;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReportRepository extends JpaRepository<Report, Integer> {
+public interface ReportRepository extends JpaRepository<Report, Long> {
 }
